requirements.txt
autopy==4.0.0
mss==6.0.0
numpy==1.19.2
opencv-python==4.4.0.42
Pillow==7.2.0
pynput==1.7.1
pyobjc-core==6.2.2
pyobjc-framework-Cocoa==6.2.2
pyobjc-framework-Quartz==6.2.2
six==1.15.0